+++ sorting_algorithms/mergesort.cpp
@@ -1,13 +1,13 @@
 extern "C"
 {
-    void merge(int arr[], int l, int m, int r)
+    void merge(double arr[], int l, int m, int r)
     {
         int i, j, k;
         int n1 = m - l + 1;
         int n2 = r - m;
 
         // Create temp arrays
-        int L[n1], R[n2];
+        double L[n1], R[n2];
 
         // Copy data to temp arrays L[] and R[]
         for (i = 0; i < n1; i++)
@@ -55,7 +55,7 @@ extern "C"
 
     // l is for left index and r is right index of the
     // sub-array of arr to be sorted
-    void mergeSort(int arr[], int l, int r)
+    void mergeSort(double arr[], int l, int r)
     {
         if (l < r)
         {
@@ -69,7 +69,7 @@ extern "C"
         }
     }
 
-    void sort(int arr[], int n) {
+    void sort(double arr[], int n) {
         mergeSort(arr, 0, n-1);
     }
 }
